pyproject.toml
[tool.poetry]
name = "fdk-atlas-dcat-service"
version = "0.1.0"
description = "A Python service that parses RDF data to JSON and serves it to other applications."
authors = ["Digdir <fellesdatakatalog@digdir.no>"]
homepage = "https://github.com/Informasjonsforvaltning/fdk-atlas-dcat-service"
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
fastapi = "~0.112.2"
fdk-rdf-parser = "^2.8.1"
python = ">=3.9, <3.12"
python-dotenv = "^1.0.1"
python-json-logger = "^2.0.7"
uvicorn = "~0.30.6"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
coverage = {extras = ["toml"], version = "^7.6.1"}
httpx = "~0.27.2"
mypy = "^1.11.2"
nox = "^2024.4.15"
nox-poetry = "^1.0.3"
pep8-naming = "^0.14.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-docker = "^3.1.1"
requests = "^2.32.3"
types-requests = "^2.32.0.20240712"
ruff = "^0.6.3"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "fdk_rdf_parser.*",
  "pythonjsonlogger.*",
  "nox",
  "nox_poetry.*",
  "requests.*",
  "httpx.*",
]

[tool.coverage.run]
source = ["fdk_atlas_dcat_service"]

[tool.coverage.report]
fail_under = 0

[tool.pytest.ini_options]
markers = [
  "unit: marks tests as unit (fast)",
  "integration: marks tests as integration (slower)",
  "contract: marks test as contract (slow)",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
